Changed app folder name
