+++ src/applications/coronavirus-vaccination-expansion/-schema.json
