+++ src/applications/coronavirus-vaccination-expansion/app-entry.jsx
