+++ src/applications/coronavirus-vaccination-expansion/config/attestation/helpers.js
