+++ src/applications/coronavirus-vaccination-expansion/helpers.js
