+++ src/applications/coronavirus-vaccination-expansion/manifest.json
