+++ src/applications/coronavirus-vaccination-expansion/reducers/index.js
