+++ src/applications/coronavirus-vaccination-expansion/sass/covid-vaccine.scss
