+++ src/applications/coronavirus-vaccination-expansion/tests/covid-vaccine.cypress.spec.js
